Correctly handling null default value properties
