+++ src/components/Launch/LaunchWorkflowForm/test/constants.ts
@@ -1,0 +1,4 @@
+export const booleanInputName = 'simpleBoolean';
+export const stringInputName = 'simpleString';
+export const stringNoLabelName = 'stringNoLabel';
+export const integerInputName = 'simpleInteger';
