+++ src/components/Launch/LaunchWorkflowForm/useLaunchWorkflowFormState.ts
@@ -1,4 +1,3 @@
-import { sortedObjectEntries } from 'common/utils';
 import { getCacheKey } from 'components/Cache';
 import { useAPIContext } from 'components/data/apiContext';
 import {
@@ -18,10 +17,7 @@ import {
 } from 'models';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { history, Routes } from 'routes';
-import {
-    defaultValueForInputType,
-    literalToInputValue
-} from './inputHelpers/inputHelpers';
+import { getInputs } from './getInputs';
 import { SearchableSelectorOption } from './SearchableSelector';
 import {
     LaunchWorkflowFormInputsRef,
@@ -30,55 +26,9 @@ import {
     ParsedInput
 } from './types';
 import {
-    formatLabelWithType,
-    getInputDefintionForLiteralType,
-    getWorkflowInputs,
     launchPlansToSearchableSelectorOptions,
     workflowsToSearchableSelectorOptions
 } from './utils';
-
-// We use a non-empty string for the description to allow display components
-// to depend on the existence of a value
-const emptyDescription = ' ';
-
-function getInputs(workflow: Workflow, launchPlan: LaunchPlan): ParsedInput[] {
-    if (!launchPlan.closure || !workflow) {
-        // TODO: is this an error?
-        return [];
-    }
-
-    const workflowInputs = getWorkflowInputs(workflow);
-    const launchPlanInputs = launchPlan.closure.expectedInputs.parameters;
-    return sortedObjectEntries(launchPlanInputs).map(value => {
-        const [name, parameter] = value;
-        const required = !!parameter.required;
-        const workflowInput = workflowInputs[name];
-        const description =
-            workflowInput && workflowInput.description
-                ? workflowInput.description
-                : emptyDescription;
-
-        const typeDefinition = getInputDefintionForLiteralType(
-            parameter.var.type
-        );
-        const typeLabel = formatLabelWithType(name, typeDefinition);
-        const label = required ? `${typeLabel}*` : typeLabel;
-
-        const defaultValue =
-            parameter.default !== undefined
-                ? literalToInputValue(typeDefinition, parameter.default)
-                : defaultValueForInputType(typeDefinition);
-
-        return {
-            defaultValue,
-            description,
-            label,
-            name,
-            required,
-            typeDefinition
-        };
-    });
-}
 
 export function useWorkflowSelectorOptions(workflows: Workflow[]) {
     return useMemo(() => {
